added jwt auth

+++ rest-api/controllers/feed.js
@@ -7,6 +7,7 @@ const { validationResult } = require('express-validator');
  * Models
  */
 const Post = require('./../models/post');
+const User = require('./../models/user');
 
 /*
  * Others
@@ -17,7 +18,7 @@ const { deleteFile } = require('./../util/file');
 const ITEMS_PER_PAGE = 2;
 
 exports.getPosts = (req, res, next) => {
-  const { query: { page: queryPage = 1 }} = req;
+  const { query: { page: queryPage = 1 } } = req;
   const page = parseFloat(queryPage);
   let totalItems;
 
@@ -42,8 +43,8 @@ exports.getPosts = (req, res, next) => {
 
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
-  const { body, file } = req;
-
+  const { body, file, userId } = req;
+  let creator;
   if (!file) {
     const error = new Error('no image sent');
     error.statusCode = 422;
@@ -59,23 +60,33 @@ exports.createPost = (req, res, next) => {
   const post = new Post({
     ...body,
     imageUrl: file.path,
-    creator: {
-      name: 'Cris'
-    }
+    creator: userId
   });
 
   post.save()
-    .then(result => {
+    .then(() => {
+      return User.findById(userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(() => {
       res
         .status(201)
         .json({
           message: 'created successfully',
-          post: result
+          post,
+          creator: {
+            _id: creator._id,
+            name: creator.name
+          },
         })
     })
     .catch(err => {
       catchError(err, next);
-    })
+    });
 }
 
 exports.getPost = (req, res, next) => {
@@ -100,7 +111,7 @@ exports.getPost = (req, res, next) => {
 exports.updatePost = (req, res, next) => {
   const errors = validationResult(req);
 
-  const { params: { postId }, body, file } = req;
+  const { params: { postId }, body, file, userId } = req;
   const { title, content, image } = body;
   let imageUrl = image;
 
@@ -125,6 +136,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== userId) {
+        const error = new Error('user not authorized to do this op');
+        error.statusCode = 403;
+        throw error;
+      }
+
       // Delete old images
       if (imageUrl !== post.imageUrl) deleteFile(post.imageUrl);
 
@@ -144,7 +162,7 @@ exports.updatePost = (req, res, next) => {
 };
 
 exports.deletePost = (req, res, next) => {
-  const { params: { postId } } = req;
+  const { params: { postId }, userId } = req;
 
   Post.findById(postId)
     .then((postDoc) => {
@@ -153,8 +171,21 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (postDoc.creator.toString() !== userId) {
+        const error = new Error('user not authorized to do this op');
+        error.statusCode = 403;
+        throw error;
+      }
+
       deleteFile(postDoc.imageUrl);
       return Post.findByIdAndRemove(postId);
+    })
+    .then(() => User.findById(userId))
+    .then(user => {
+      /** Deleting post related on user models */
+      user.posts.pull(postId);
+      return user.save();
     })
     .then(() => {
       res
